% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyfunctions.R
\name{declarativeApply}
\alias{declarativeApply}
\alias{\%>lapply\%}
\alias{\%>sapply\%}
\title{Declarative Apply}
\usage{
x \%>lapply\% y

x \%>sapply\% y
}
\arguments{
\item{x}{input list}

\item{y}{function}
}
\description{
Apply functions over lists
}
\examples{

# Input:
input <- list(a = c(1,2,3),       
              b = c(4,4,5), 
              c = c(1,2,3,4,5))


# Chain lapply and sapply:
input \%>lapply\% sum \%>sapply\% sqrt

}
