% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyfunctions.R
\name{declarativeR_Apply}
\alias{declarativeR_Apply}
\alias{declapply}
\title{Apply functions over lists}
\arguments{
\item{x}{}

\item{y}{function definition}
}
\description{
Apply functions over lists
}
