% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setfunctions.R
\name{\%-\%}
\alias{\%-\%}
\title{Set difference}
\usage{
x \%-\% y
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
All elements in x that are not in y.
}
\description{
Set difference
}
\examples{
 
 
}
