% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setfunctions.R
\name{\%intersect\%}
\alias{\%intersect\%}
\title{Intersect}
\usage{
x \%intersect\% y
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
All unique elements that are in both x and y.
}
\description{
Intersect
}
