% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringfunctions.R
\name{\%+\%}
\alias{\%+\%}
\title{String concatenation operator}
\usage{
x \%+\% y
}
\arguments{
\item{x}{string or vector of strings}

\item{y}{string or vector of strings}
}
\value{
concatenation; elements are recycled.
}
\description{
String concatenation operator
}
\examples{

# Simple concatenation: 
"a" \%+\% "b"

# Pairwise concatenation: 
c("a","b") \%+\% c("a")

c("a","b") \%+\% c("a","b") 

# Null on one side:
"a" \%+\%
}
