% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyfunctions.R
\name{declarativeApply}
\alias{declarativeApply}
\alias{\\%>lapply\\%}
\alias{\%>lapply\%}
\alias{\\%>sapply\\%}
\alias{\%>sapply\%}
\title{test}
\usage{
x \%>lapply\% y

x \%>sapply\% y
}
\arguments{
\item{x}{input list}

\item{y}{function}
}
\description{
Apply functions over lists
}
